refactor(posts): move PostNavigator to aggregate module
