+++ features/posts/aggregate/presentation/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/aggregate/presentation/navigation/PostNavigator.kt
@@ -1,10 +1,18 @@
 package com.parksupark.soomjae.features.posts.aggregate.presentation.navigation
 
 import androidx.navigation.NavHostController
+import com.parksupark.soomjae.core.presentation.ui.navigation.SoomjaeNavigator
 import com.parksupark.soomjae.core.presentation.ui.navigation.overridden
-import com.parksupark.soomjae.features.posts.common.presentation.navigation.PostNavigator
+import com.parksupark.soomjae.features.posts.community.presentation.navigation.CommunityNavigator
+
+interface PostNavigator : SoomjaeNavigator, CommunityNavigator {
+    override fun navigateToCommunityWrite()
+
+    override fun navigateToCommunityDetail(postId: Long)
+}
 
 private class SoomjaePostNavigator(override val navController: NavHostController) : PostNavigator {
+
     override fun navigateBack() {
         overridden()
     }
