+++ features/posts/aggregate/presentation/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/aggregate/presentation/post/PostRoute.kt
@@ -3,8 +3,8 @@ package com.parksupark.soomjae.features.posts.aggregate.presentation.post
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import com.parksupark.soomjae.features.posts.aggregate.presentation.navigation.PostNavigator
 import com.parksupark.soomjae.features.posts.common.presentation.PostAction
-import com.parksupark.soomjae.features.posts.common.presentation.navigation.PostNavigator
 
 @Composable
 internal fun PostRoute(
