--- features/posts/common/presentation/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/common/presentation/navigation/PostNavigator.kt
@@ -1,9 +1,0 @@
-package com.parksupark.soomjae.features.posts.common.presentation.navigation
-
-import com.parksupark.soomjae.core.presentation.ui.navigation.SoomjaeNavigator
-
-interface PostNavigator : SoomjaeNavigator {
-    fun navigateToCommunityWrite()
-
-    fun navigateToCommunityDetail(postId: Long)
-}
